[Bronze II] Title: 숫자, Time: 24 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -2,20 +2,20 @@
 using namespace std;
 
 int main() {
-    int A, B;
-    int tmp;
+    long long A, B;
+    long long tmp;
     cin >> A >> B;
 
     if (A != B) cout << abs(A-B)-1 << '\n';
     else cout << 0 << '\n';
-    
+
     if (A > B) {
         tmp = B;
         B = A;
         A = tmp;
     }
 
-    for (int i = A+1; i < B; i++) {
+    for (long long i = A+1; i < B; i++) {
         cout << i << " ";
     }
 }
